alchemist/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pytesseract==0.3.13
Pillow==11.3.0
pdfplumber==0.11.7
markdown==3.8.2
jinja2==3.1.6
matplotlib==3.10.5
numpy==2.3.2
